Add auth + tests to `ListCertificates`
